+++ src/alert.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+extern void alert();
