+++ src/main.cpp
@@ -1,5 +1,6 @@
 ﻿// warning: shitty code incoming
 // good luck reading it, i should probably split it up into multiple files
+// i got u - coop
 #include <Geode/Geode.hpp>
 #include <Geode/modify/CCSprite.hpp>
 #include <Geode/modify/MenuLayer.hpp>
@@ -7,27 +8,17 @@
 #include <Geode/modify/GJGarageLayer.hpp>
 #include <Geode/modify/LevelBrowserLayer.hpp>
 #include <Geode/modify/GJGroundLayer.hpp>
+#include "./alert.hpp"
 
 #include <random>
 // setup random stuff
 std::random_device dev;
 std::mt19937 rng(dev());
 std::uniform_int_distribution<std::mt19937::result_type> distBG(1, 59);
-std::uniform_int_distribution<std::mt19937::result_type> distGround(1, 22);
-std::uniform_int_distribution<std::mt19937::result_type> distGroundLine(1, 3);
 
 using namespace geode::prelude;
 
-std::unordered_set<std::string> cornerSpriteNames = {
-    "GJ_sideArt_001.png",
-    "dailyLevelCorner_001.png",
-    "gauntletCorner_001.png",
-    "rewardCorner_001.png",
-    "treasureRoomSpiderweb_001.png"
-};
-
-void alert(std::string labelName, std::string taskName, cocos2d::CCNode* _this)
-{
+void alert(std::string labelName, std::string taskName, cocos2d::CCNode* _this) {
     std::cout << "couldn't find " + labelName + " for " + taskName << std::endl;
 
     if (Mod::get()->getSettingValue<bool>("suppress-warnings")) return;
@@ -42,62 +33,6 @@ void alert(std::string labelName, std::string taskName, cocos2d::CCNode* _this)
 }
 
 bool loadingMainMenu = false;
-
-// "randomise-main-menu-bg" (ground half)
-class $modify(GJGroundLayer)
-{
-    static GJGroundLayer* create(int groundType, int lineType)
-    {
-        if (!loadingMainMenu) return GJGroundLayer::create(groundType, lineType);
-
-        int64_t ground = Mod::get()->getSettingValue<int64_t>("force-ground");
-
-        //std::cout << ground << std::endl;
-
-        if (ground != 0)
-        {
-            //std::cout << "yah" << std::endl;
-            groundType = std::stoi(std::to_string(ground));
-            //std::cout << "did that" << std::endl;
-        }
-        else if (Mod::get()->getSettingValue<bool>("randomise-main-menu-bg"))
-        {
-            groundType = distGround(rng);
-            lineType = distGroundLine(rng);
-        }
-
-        //std::cout << "abt to create" << std::endl;
-        auto* groundLayer = GJGroundLayer::create(groundType, lineType);
-        //std::cout << "created!" << std::endl;
-
-        if (Mod::get()->getSettingValue<bool>("remove-ground"))
-            groundLayer->setVisible(false);
-
-        return groundLayer;
-    }
-};
-
-// I may or may not have ~~stolen~~ kindly borrowed some of this from NoControllerGlyphs
-// it's pretty clever tbh
-// "remove-corners" + "remove-level-select-top"
-class $modify(CCSprite) {
-    static void onModify(auto & self)
-    {
-        self.setHookPriority("CCSprite::createWithSpriteFrameName", 1000);
-    }
-
-    static CCSprite* createWithSpriteFrameName(char const* frameName) {
-        CCSprite* sprite = CCSprite::createWithSpriteFrameName(frameName);
-
-        if (cornerSpriteNames.find(frameName) != cornerSpriteNames.end() && Mod::get()->getSettingValue<bool>("remove-corners"))
-            sprite->setVisible(false); // remove corner
-
-        if (strcmp(frameName, "GJ_topBar_001.png") == 0 && Mod::get()->getSettingValue<bool>("remove-level-select-top"))
-            sprite->setVisible(false); // remove level select topper thingymabob
-
-        return sprite;
-    }
-};
 
 // "title-text" + "fix-main-menu-settings-gamepad" + "title-buttons" + "replace-more-games-w-texture" + "randomise-main-menu-bg" (bg half) + "move-player-to-corner"
 class $modify(MenuLayer)
@@ -506,271 +441,5 @@ class $modify(MenuLayer)
         if (!Mod::get()->getSettingValue<bool>("fire-in-the-hole")) return;
 
         FMODAudioEngine::sharedEngine()->playEffect("fire.ogg"_spr);
-    }
-};
-
-
-// "map-packs" + "move-betterinfo" + "map-packs-hof"
-class $modify(MyCreatorLayer, CreatorLayer)
-{
-    static void onModify(auto & self)
-    {
-        // run after betterinfo
-        self.setHookPriority("CreatorLayer::init", -100);
-    }
-
-    // definitely not taken from BetterInfo :) (thanks cvolton)
-    void onFame(CCObject* object)
-    {
-        auto searchObject = GJSearchObject::create(SearchType::HallOfFame);
-        auto browserLayer = LevelBrowserLayer::scene(searchObject);
-
-        auto transitionFade = CCTransitionFade::create(0.5, browserLayer);
-
-        CCDirector::sharedDirector()->pushScene(transitionFade);
-    }
-
-    bool init()
-    {
-        if (!CreatorLayer::init()) return false;
-
-        if (Mod::get()->getSettingValue<bool>("map-packs"))
-        {
-            if (Mod::get()->getSettingValue<bool>("map-packs-hof"))
-            {
-                if (Mod::get()->getSettingValue<bool>("suppress-warnings")) goto endOfMapPacks;
-
-                auto alert = FLAlertLayer::create(
-                    "Notice (from GDTweaks)",
-                    "GDTweaks couldn't <cy>remove the map pack button</c> because <cj>\"Replace \"Map Packs\" with Hall Of Fame\"</c> is enabled. You can suppress these dialogs in the GDTweaks settings.",
-                    "OK"
-                );
-                alert->m_scene = this;
-                alert->show();
-
-                goto endOfMapPacks;
-            }
-
-            std::string buttonChecks[6] = {
-                "daily-button",
-                "weekly-button",
-                "event-button",
-                "gauntlets-button",
-                "map-button",
-                "map-packs-button"
-            };
-
-            // firstly we have to check every single thing is here - if not then a mod
-            // will have incompatibilities
-            for (int i = 0; i < 6; i++)
-            {
-                std::string labelName = buttonChecks[i];
-                if (!this->getChildByIDRecursive(labelName))
-                {
-                    alert(labelName, "remove the map pack button", this);
-                    goto endOfMapPacks;
-                }
-            }
-
-            // guaranteed to exist bc of the check above
-            auto mpb = this->getChildByIDRecursive("map-packs-button");
-
-            // move the map to where map packs used to be
-            this->getChildByIDRecursive("map-button")->setPosition(mpb->getPosition());
-
-            // then shift all the rest over
-            // somehow these magic numbers work fine on every aspect ratio I've tested (4:3 (actually 3:2 but whatever), 16:9 and 21:9 (actually 20:9 but ignore my weird aspect ratio issues ok?))
-            this->getChildByIDRecursive("daily-button")->setPositionX(79.35);
-            this->getChildByIDRecursive("weekly-button")->setPositionX(176.45);
-            this->getChildByIDRecursive("event-button")->setPositionX(273.55);
-            this->getChildByIDRecursive("gauntlets-button")->setPositionX(370.65);
-
-            // and remove map pack button
-            mpb->removeFromParent();
-        }
-
-    endOfMapPacks:
-
-        if (Mod::get()->getSettingValue<bool>("move-betterinfo"))
-        {
-            if (!Loader::get()->isModLoaded("cvolton.betterinfo")) goto endOfBetterInfo;
-
-            // move betterinfo button to bottom left (if it exists)
-            // for some reason setPosition doesn't seem to work for me
-            // so two setPositionX/Ys will have to do
-            if (auto biButton = this->getChildByID("cvolton.betterinfo/center-right-menu"))
-            {
-                biButton->setPositionX(28);
-                biButton->setPositionY(28);
-            }
-            else
-            {
-                alert("cvolton.betterinfo/center-right-menu", "move the betterinfo button to the corner", this);
-                goto endOfBetterInfo; // not needed but might as well in case expansion is needed
-            }
-        }
-
-    endOfBetterInfo:
-
-        if (Mod::get()->getSettingValue<bool>("map-packs-hof"))
-        {
-            auto mpb = this->getChildByIDRecursive("map-packs-button");
-            if (!mpb)
-            {
-                alert("map-packs-button", "replacing the Map Packs button with the Hall Of Fame", this);
-                goto endOfHallOfFame;
-            }
-            mpb->setVisible(false);
-            auto mapSprite = typeinfo_cast<CCSprite*>(mpb->getChildren()->objectAtIndex(0));
-
-            // taken from BetterInfo :) thanks cvolton
-            auto* spr = CCSprite::create("GJ_fameBtn_001.png"_spr);
-            auto fameBtn = CCMenuItemSpriteExtra::create(
-                spr,
-                this,
-                menu_selector(MyCreatorLayer::onFame)
-            );
-            // i could just set it to "fame-button" but im nice :)
-            fameBtn->setID("undefined0.gdtweaks/fame-button");
-
-            auto cbm = this->getChildByIDRecursive("creator-buttons-menu");
-            // this should exist in theory but might as well check
-            if (!cbm)
-            {
-                alert("creator-buttons-menu", "replacing the Map Packs button with the Hall Of Fame", this);
-                goto endOfHallOfFame;
-            }
-
-            // add the fame button
-            cbm->addChild(fameBtn);
-            spr->setScale(0.8); // for some reason it's just so much bigger than everything else without this
-
-            // for some reason i cant get the positioning or anything right so ill just set everything to what
-            // the map packs had it at
-            auto mapSpriteContentSize = mapSprite->getContentSize();
-            mapSpriteContentSize.width = 93.5;
-            spr->setContentSize(mapSpriteContentSize);
-            fameBtn->setContentSize(mpb->getContentSize());
-            spr->setPosition(mapSprite->getPosition());
-            fameBtn->setPosition(mpb->getPosition());
-        }
-
-    endOfHallOfFame:
-
-        // this is just a nice thing to do
-        if (auto biButton = this->getChildByID("cvolton.betterinfo/center-right-menu")) biButton->setZOrder(1);
-
-        return true;
-    }
-};
-
-// "garage-bg-remove" + "garage-lock-tap-remove" + 
-class $modify(GJGarageLayer)
-{
-    static void onModify(auto & self)
-    {
-        self.setHookPriority("GJGarageLayer::init", 100);
-    }
-
-    bool init()
-    {
-        if (!GJGarageLayer::init()) return false;
-
-        if (Mod::get()->getSettingValue<bool>("garage-bg-remove"))
-        {
-            // remove the background from the icon kit
-            for (int i = 0; i < this->getChildrenCount(); i++)
-            {
-                if (auto bg = typeinfo_cast<CCScale9Sprite*>(this->getChildren()->objectAtIndex(i)))
-                {
-                    bg->setVisible(false);
-                    goto endOfGarageBGRemove;
-                }
-            }
-
-            //std::cout << "couldnt find ccscale9sprite for icon kit" << std::endl;
-            alert("icon kit background", "removing the icon kit background", this);
-        }
-
-    endOfGarageBGRemove:
-
-        if (Mod::get()->getSettingValue<bool>("garage-lock-tap-remove"))
-        {
-            // remove tap for more info lock
-
-            // lock->setVisible(false);
-            // player icon -> setpositiony 222
-            // icon type selector - > setposition y 171
-            // weird ground line -> psoiton y 197
-
-            if (auto lock = typeinfo_cast<CCSprite*>(this->getChildByIDRecursive("tap-more-hint")))
-                lock->setVisible(false);
-            else
-            {
-                alert("tap-more-hint", "removing the tap for more info lock", this);
-                goto endOfGarageLockTapRemove;
-            }
-
-            if (auto player = typeinfo_cast<SimplePlayer*>(this->getChildByIDRecursive("player-icon")))
-                player->setPositionY(222);
-            else
-            {
-                alert("player-icon", "removing the tap for more info lock", this);
-                goto endOfGarageLockTapRemove;
-            }
-
-            if (auto iconTypeSelector = typeinfo_cast<CCMenu*>(this->getChildByIDRecursive("category-menu")))
-                iconTypeSelector->setPositionY(171);
-            else
-            {
-                alert("category-menu", "removing the tap for more info lock", this);
-                goto endOfGarageLockTapRemove;
-            }
-
-            if (auto line = typeinfo_cast<CCSprite*>(this->getChildByIDRecursive("floor-line")))
-                line->setPositionY(197);
-            else
-            {
-                alert("floor-line", "removing the tap for more info lock", this);
-                goto endOfGarageLockTapRemove;
-            }
-        }
-
-    endOfGarageLockTapRemove:
-
-        return true;
-    }
-};
-
-// "fix-joystick-scroll"
-class $modify(LevelBrowserLayer)
-{
-    bool init(GJSearchObject * p0)
-    {
-        if (!LevelBrowserLayer::init(p0)) return false;
-
-        // put joystick controller stuff in correct position
-        if (Mod::get()->getSettingValue<bool>("fix-joystick-scroll"))
-        {
-            auto nodes = this->getChildren();
-            for (int i = 0; i < this->getChildrenCount(); ++i)
-            {
-                auto node = nodes->objectAtIndex(i);
-
-                if (auto spriteNode = typeinfo_cast<CCSprite*>(node))
-                {
-                    // test if it's in the correct position to be the controller buttons
-                    // there may be a better way to test this but not that I know of
-                    if (spriteNode->getPositionX() == 545)
-                    {
-                        // if so then fix it
-                        spriteNode->setPositionY(spriteNode->getPositionY() - 140);
-                        spriteNode->setPositionX(30);
-                    }
-                }
-            }
-        }
-
-        return true;
     }
 };
